+++ src/actions/upgrade-action.ts
@@ -2,6 +2,7 @@ import type {AppendKeyValue} from '@helpers/type';
 
 import fs from 'node:fs';
 
+import {catchPnpmExec} from '@helpers/actions/upgrade/catch-pnpm-exec';
 import {getBetaVersion} from '@helpers/beta';
 import {getCanaryVersion} from '@helpers/canary';
 import {checkIllegalComponents, getConditionLatestVersion} from '@helpers/check';
@@ -13,9 +14,9 @@ import {getPackageInfo} from '@helpers/package';
 import {setupPnpm} from '@helpers/setup';
 import {upgrade, writeUpgradeVersion} from '@helpers/upgrade';
 import {getColorVersion, getPackageManagerInfo, transformPeerVersion} from '@helpers/utils';
-import {type NextUIComponents} from 'src/constants/component';
+import {type HeroUIComponents} from 'src/constants/component';
 import {resolver} from 'src/constants/path';
-import {NEXT_UI} from 'src/constants/required';
+import {HERO_UI} from 'src/constants/required';
 import {store} from 'src/constants/store';
 import {getAutocompleteMultiselect, getMultiselect, getSelect} from 'src/prompts';
 import {compareVersions, getLatestVersion} from 'src/scripts/helpers';
@@ -32,7 +33,7 @@ interface UpgradeActionOptions {
 }
 
 type TransformComponent = Required<
-  AppendKeyValue<NextUIComponents[0], 'latestVersion', string> & {isLatest: boolean}
+  AppendKeyValue<HeroUIComponents[0], 'latestVersion', string> & {isLatest: boolean}
 >;
 
 function extraCompareVersions(
@@ -73,14 +74,14 @@ export async function upgradeAction(components: string[], options: UpgradeAction
   const {allDependencies, currentComponents, dependencies, devDependencies, packageJson} =
     getPackageInfo(packagePath, false);
 
-  const isNextUIAll = !!allDependencies[NEXT_UI];
+  const isHeroUIAll = !!allDependencies[HERO_UI];
 
   const transformComponents: TransformComponent[] = [];
 
   await Promise.all(
     currentComponents.map(async (component) => {
       const latestVersion =
-        store.nextUIComponentsMap[component.name]?.version ||
+        store.heroUIComponentsMap[component.name]?.version ||
         (await getLatestVersion(component.package));
       const mergedVersion = beta
         ? await getBetaVersion(component.package)
@@ -97,9 +98,9 @@ export async function upgradeAction(components: string[], options: UpgradeAction
     })
   );
 
-  // If no Installed NextUI components then exit
-  if (!transformComponents.length && !isNextUIAll) {
-    Logger.prefix('error', `No NextUI components detected in your package.json at: ${packagePath}`);
+  // If no Installed HeroUI components then exit
+  if (!transformComponents.length && !isHeroUIAll) {
+    Logger.prefix('error', `No HeroUI components detected in your package.json at: ${packagePath}`);
 
     return;
   }
@@ -107,23 +108,23 @@ export async function upgradeAction(components: string[], options: UpgradeAction
   if (all) {
     components = currentComponents.map((component) => component.package);
   } else if (!components.length) {
-    // If have the main nextui then add
-    if (isNextUIAll) {
-      const version = transformPeerVersion(allDependencies[NEXT_UI]);
+    // If have the main heroui then add
+    if (isHeroUIAll) {
+      const version = transformPeerVersion(allDependencies[HERO_UI]);
       const latestVersion = getConditionLatestVersion(store.beta, store.canary);
-      const nextuiData = {
+      const herouiData = {
         isLatest: extraCompareVersions(version, latestVersion, store.beta, store.canary),
         latestVersion,
-        package: NEXT_UI,
+        package: HERO_UI,
         version
       } as TransformComponent;
 
-      transformComponents.push(nextuiData);
+      transformComponents.push(herouiData);
     }
 
     // If all package is latest then pass
     if (transformComponents.every((component) => component.isLatest)) {
-      Logger.success('✅ All NextUI packages are up to date');
+      Logger.success('✅ All HeroUI packages are up to date');
       process.exit(0);
     }
 
@@ -160,8 +161,8 @@ export async function upgradeAction(components: string[], options: UpgradeAction
   }
 
   components = components.map((c) => {
-    if (store.nextUIComponentsMap[c]?.package) {
-      return store.nextUIComponentsMap[c]!.package;
+    if (store.heroUIComponentsMap[c]?.package) {
+      return store.heroUIComponentsMap[c]!.package;
     }
 
     return c;
@@ -173,7 +174,7 @@ export async function upgradeAction(components: string[], options: UpgradeAction
   let result = await upgrade({
     all,
     allDependencies,
-    isNextUIAll,
+    isHeroUIAll,
     upgradeOptionList
   });
   let ignoreList: string[] = [];
@@ -239,18 +240,20 @@ export async function upgradeAction(components: string[], options: UpgradeAction
         upgradePackageList: result
       });
 
-      fs.writeFileSync(packagePath, JSON.stringify(packageJson, null, 2));
+      fs.writeFileSync(packagePath, JSON.stringify(packageJson, null, 2), 'utf-8');
 
       Logger.newLine();
       Logger.success('✅ Upgrade version written to package.json');
       process.exit(0);
     } else {
-      await exec(
-        `${packageManager} ${install} ${result.reduce((acc, component, index) => {
-          return `${acc}${index === 0 ? '' : ' '}${
-            component.package
-          }@${component.latestVersion.replace(colorMatchRegex, '')}`;
-        }, '')}`
+      await catchPnpmExec(() =>
+        exec(
+          `${packageManager} ${install} ${result.reduce((acc, component, index) => {
+            return `${acc}${index === 0 ? '' : ' '}${
+              component.package
+            }@${component.latestVersion.replace(colorMatchRegex, '')}`;
+          }, '')}`
+        )
       );
     }
   }