+++ src/helpers/debug.ts
@@ -2,6 +2,7 @@ import {writeFileSync} from 'node:fs';
 
 import {getStoreSync, store} from 'src/constants/store';
 
+import {catchPnpmExec} from './actions/upgrade/catch-pnpm-exec';
 import {exec} from './exec';
 import {Logger} from './logger';
 import {getPackageInfo} from './package';
@@ -12,7 +13,7 @@ export async function debugExecAddAction(cmd: string, components: string[] = [])
       Logger.warn(`Debug: ${component}`);
     }
   } else {
-    await exec(cmd);
+    await catchPnpmExec(() => exec(cmd));
   }
 }
 
@@ -22,7 +23,7 @@ export function debugAddedPkg(components: string[], packagePath: string) {
   const {dependencies, packageJson} = getPackageInfo(packagePath);
 
   for (const component of components) {
-    const compData = store.nextUIComponentsMap[component];
+    const compData = store.heroUIComponentsMap[component];
 
     if (!compData) continue;
 
@@ -37,7 +38,8 @@ export function debugAddedPkg(components: string[], packagePath: string) {
       },
       null,
       2
-    )
+    ),
+    'utf-8'
   );
 }
 
@@ -47,7 +49,7 @@ export function debugRemovedPkg(components: string[], packagePath: string) {
   const {dependencies, packageJson} = getPackageInfo(packagePath);
 
   for (const component of components) {
-    const compData = store.nextUIComponentsMap[component];
+    const compData = store.heroUIComponentsMap[component];
 
     if (!compData) continue;
     delete dependencies[compData.package];
@@ -61,6 +63,7 @@ export function debugRemovedPkg(components: string[], packagePath: string) {
       },
       null,
       2
-    )
+    ),
+    'utf-8'
   );
 }