+++ src/helpers/type.ts
@@ -1,6 +1,6 @@
 /* eslint-disable @typescript-eslint/no-explicit-any */
 
-import type {NextUIComponentsMap, StoreKeys} from 'src/constants/store';
+import type {HeroUIComponentsMap, StoreKeys} from 'src/constants/store';
 import type {Components} from 'src/scripts/helpers';
 
 /**
@@ -33,6 +33,15 @@ export type RequiredKey<T, Key extends keyof T> = {
   [K in keyof T as K extends Key ? never : K]?: T[K];
 } & {
   [K in Key]-?: T[K];
+};
+
+/**
+ * @example PartialKey<{a: 1, b: 2}, a> => {a?: 1, b: 2}
+ */
+export type PartialKey<T, Key extends keyof T> = {
+  [K in keyof T as K extends Key ? never : K]: T[K];
+} & {
+  [K in Key]?: T[K];
 };
 
 export type ChalkColor =
@@ -75,16 +84,16 @@ export type ChalkColor =
 
 export type ExtractStoreData<T extends StoreKeys> = T extends 'latestVersion' | 'cliLatestVersion'
   ? string
-  : T extends 'nextUIComponents'
+  : T extends 'heroUIComponents'
     ? Components
-    : T extends 'nextUIComponentsKeys' | 'nextUIcomponentsPackages'
+    : T extends 'heroUIComponentsKeys' | 'heroUIcomponentsPackages'
       ? string[]
-      : T extends 'nextUIComponentsKeysSet'
+      : T extends 'heroUIComponentsKeysSet'
         ? Set<string>
-        : T extends 'nextUIComponentsMap'
-          ? NextUIComponentsMap
-          : T extends 'nextUIComponentsPackageMap'
-            ? NextUIComponentsMap
+        : T extends 'heroUIComponentsMap'
+          ? HeroUIComponentsMap
+          : T extends 'heroUIComponentsPackageMap'
+            ? HeroUIComponentsMap
             : never;
 
 /**